rcu: Expand SRCU ->completed to 64 bits
When rcutorture used only the low-order 32 bits of the grace-period
number, it was not a problem for SRCU to use a 32-bit completed field.
However, rcutorture now uses the full 64 bits on 64-bit systems, so
this commit converts SRCU's ->completed field to unsigned long so as to
provide 64 bits on 64-bit systems.

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Ad!thya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Mrinal Ghosh <mg712702@gmail.com>