rcu: Remove redundant callback-list initialization
The RCU callback lists are initialized in both rcu_boot_init_percpu_data()
and rcu_init_percpu_data().  The former is intended for initializing
immutable data, so this commit removes the initialization from
rcu_boot_init_percpu_data() and leaves it in rcu_init_percpu_data().
This change prepares for permitting callbacks to be queued very early
in boot.

Change-Id: I0d4b1395e1283819bbb8a9835ba77ec6a86960ae
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>
Signed-off-by: Ad!thya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Mrinal Ghosh <mg712702@gmail.com>