staging: LMK: Evaluate regular swap space when calculating memory
Android lowmemorykiller was evaluating only swapcache pages, but
not "real" linux swap space: this was making it to usually kill
processes anyway when the kernel swappiness parameter was high.

Include swap space(s) evaluation in the vmpressure and shrink
algorithms to allow swap space to be used as a last resort.

Signed-off-by: Mrinal Ghosh <mg712702@gmail.com>